APP - changed place rating type from double for num because i had a problem with parse from json, Fixed place rating stars display when rating there was integer without comma

+++ awesome_places/lib/src/features/explore/presentation/place_details_screen.dart
@@ -119,7 +119,7 @@ class _PlaceDetailsScreenState extends State<PlaceDetailsScreen> {
                           ),
                         ),
                         const SizedBox(width: 7),
-                        Stars(rating: widget.place.rating),
+                        Stars(rating: widget.place.rating.toDouble()),
                       ],
                     ),
                     const PlaceDivider(),
